fix: Refatorando funcionalidades de característica: Corte de Cabelo

+++ src/modules/caracteristicasPaciente/corteCabelo/useCases/deleteCorteCabelo/DeleteCorteCabeloUseCase.ts
@@ -2,6 +2,7 @@ import { ICorteCabeloRepository } from '@modules/caracteristicasPaciente/corteCa
 import { inject, injectable } from 'tsyringe';
 
 import { AppError } from '@shared/errors/AppError';
+import { Messages } from '@shared/messages/Messages';
 
 @injectable()
 class DeleteCorteCabeloUseCase {
@@ -11,9 +12,15 @@ class DeleteCorteCabeloUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    if (!id) {
+      throw new AppError(
+        `${Messages.MISSING_PARAMETERS}: ID de Corte do Cabelo`,
+      );
+    }
+
     const corteCabelo = await this.corteCabeloRepository.listById(id);
     if (!corteCabelo) {
-      throw new AppError('CorteCabelo not found', 404);
+      throw new AppError(Messages.CHARACTERISTICS_NOT_FOUND, 404);
     }
 
     await this.corteCabeloRepository.delete(id);
